# src/trapo-server/openapi/trapo.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trapo Server API",
    "description": "Rust Axum API for Trapo OCR workbench.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "0.1.5"
  },
  "paths": {
    "/api/diagnostics/analytics": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "operationId": "diagnostics_analytics_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticAnalyticsPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostics/models": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "operationId": "diagnostics_models_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticModelsPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostics/progress": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "operationId": "diagnostics_progress_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticProgressPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostics/runs": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "operationId": "diagnostics_runs_doc",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticRunsPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/diagnostics/trace": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "operationId": "diagnostics_trace_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file_hash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_no",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosticTracePayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "list_documents_doc",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{file_hash}": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "get_document_doc",
        "parameters": [
          {
            "name": "file_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{file_hash}/preview-images": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "preview_images_doc",
        "parameters": [
          {
            "name": "file_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewImagesPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{file_hash}/preview-images/{variant}/{page_no}": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "preview_image_doc",
        "parameters": [
          {
            "name": "file_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variant",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preview image bytes",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{file_hash}/regions": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "document_regions_doc",
        "parameters": [
          {
            "name": "file_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRegionsPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{file_hash}/regions/{region_id}/snippet": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "region_snippet_doc",
        "parameters": [
          {
            "name": "file_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "region_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Region image snippet bytes",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{file_hash}/text": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "document_text_doc",
        "parameters": [
          {
            "name": "file_hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentTextPayload"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ingest/runs/{run_id}": {
      "get": {
        "tags": [
          "ingest"
        ],
        "operationId": "get_run_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestRunRecord"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/runs/{run_id}/events": {
      "get": {
        "tags": [
          "ingest"
        ],
        "operationId": "run_events_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent run snapshots",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/ingest/runs/{run_id}/stop": {
      "post": {
        "tags": [
          "ingest"
        ],
        "operationId": "stop_run_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Stop requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestRunRecord"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/start": {
      "post": {
        "tags": [
          "ingest"
        ],
        "operationId": "start_ingest_doc",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Ingest run queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestStartResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/recent": {
      "get": {
        "tags": [
          "logs"
        ],
        "operationId": "logs_doc",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogsPayload"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/models/{model_id}/cancel": {
      "post": {
        "tags": [
          "models"
        ],
        "operationId": "cancel_model_doc",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Download cancellation requested",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownloadRecord"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/download": {
      "post": {
        "tags": [
          "models"
        ],
        "operationId": "download_model_doc",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDownloadRecord"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/events": {
      "get": {
        "tags": [
          "models"
        ],
        "operationId": "model_events_doc",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent model download progress events",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/{model_id}/select": {
      "post": {
        "tags": [
          "models"
        ],
        "operationId": "select_model_doc",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Model selected for subsequent ingest runs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSelectRecord"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/ocr/events": {
      "get": {
        "tags": [
          "diagnostics"
        ],
        "operationId": "ocr_events_doc",
        "parameters": [
          {
            "name": "run_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "file_hash",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page_no",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "since_sequence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrReplayPayload"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "search_documents_doc",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DuckDB-backed document text search",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentsPayload"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "operationId": "settings_doc",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsPayload"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "operationId": "update_settings_doc",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsPayload"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DiagnosticAnalyticsPayload": {
        "type": "object",
        "required": [
          "summary",
          "by_pipeline_step",
          "by_category",
          "slow_spans",
          "recommendations"
        ],
        "properties": {
          "by_category": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticBreakdownRecord"
            }
          },
          "by_pipeline_step": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticBreakdownRecord"
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticRecommendationRecord"
            }
          },
          "slow_spans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticSlowSpanRecord"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/DiagnosticAnalyticsSummary"
          }
        }
      },
...
      "DiagnosticModelsPayload": {
        "type": "object",
        "required": [
          "model_leases"
        ],
        "properties": {
          "model_leases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticModelLeaseRecord"
            }
          }
        }
      },
...
      "DiagnosticRunsPayload": {
        "type": "object",
        "required": [
          "runs"
        ],
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiagnosticRunRecord"
            }
          }
        }
      },
...
      "ErrorPayload": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}